Refactor UI item info classes; add attribute system
Separated PdmUiItemInfo and PdmOptionItemInfo into their own source/header files for better modularity. Updated CMakeLists.txt accordingly. Enhanced PdmUiItem with a flexible map-based attribute system supporting per-config and default attributes, including new set/get and attributeNames methods.
